Organize files according to MVC. Refactoring
